refactor: déplace InitScreen.jsx → features/init/
